+++ tests/test_api_jwt.py
@@ -142,7 +142,7 @@ class TestJWT:
                        'eyJpYXQiOiJub3QtYW4taW50In0.'
                        'H1GmcQgSySa5LOKYbzGm--b1OmRbHFkyk8pq811FzZM')
 
-        with pytest.raises(DecodeError):
+        with pytest.raises(InvalidIssuedAtError):
             jwt.decode(example_jwt, 'secret')
 
     def test_decode_raises_exception_if_nbf_is_not_int(self, jwt):
@@ -153,13 +153,6 @@ class TestJWT:
 
         with pytest.raises(DecodeError):
             jwt.decode(example_jwt, 'secret')
-
-    def test_decode_raises_exception_if_iat_in_the_future(self, jwt):
-        now = datetime.utcnow()
-        token = jwt.encode({'iat': now + timedelta(days=1)}, key='secret')
-
-        with pytest.raises(InvalidIssuedAtError):
-            jwt.decode(token, 'secret')
 
     def test_encode_datetime(self, jwt):
         secret = 'secret'